refactor: use kebab-case filenames
