+++ src/models/challenge.ts
