+++ src/models/deal.ts
