+++ src/models/reward-catalog.ts
