+++ src/models/training.ts
