+++ src/routes/me.ts
@@ -1,6 +1,6 @@
 import { Router, Request, Response } from 'express';
 import auth from '../middleware/auth';
-import User from '../models/User';
+import User from '../models/user';
 
 const router = Router();
